child class for additional controllers
